OSW-5 Setup Development Environment Update folder structure, move some folders to src Rename folders to plural: "lib" to "libs"... Done install Cyrpess (E2E only) and Jest Add Logo HCMUS (placeholder for assets folder) Update readme file with new package and folder structure
