+++ src/app/api/payment/payos/payment-requests/route.ts
@@ -1,7 +1,7 @@
 // Configuration source: https://vercel.com/docs/storage/vercel-postgres/quickstart
 import { NextResponse } from 'next/server';
-import { PayOSObject } from '@interface/PayOSObject'
-import { getPayOSSignature, generateOrderCode, getUnixTimeStamp, calcTotalPrice } from '@utils/payOSUtils';
+import { PayOSObject } from '@/interfaces/PayOSObject';
+import { getPayOSSignature, generateOrderCode, getUnixTimeStamp, calcTotalPrice } from '@/utils/payOSUtils';
 
 // Do not use `export default` here, as it will break the routing
 export async function GET(req: Request) {
