+++ src/interfaces/PayOSObject.ts
