+++ src/stories/button.stories.tsx
@@ -2,7 +2,7 @@ import type { Meta, StoryObj } from '@storybook/react';
 
 import * as React from 'react';
 import { Button, ButtonProps } from '@mui/material';
-import { MyButton } from '../lib/ui/components/Button';
+import { MyButton } from '@/libs/ui/components/Button';
 
 const meta: Meta = {
   component: MyButton,
