+++ src/stories/button2.stories.tsx
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from '@storybook/react';
 
 import * as React from 'react';
-import { Button, ButtonProps } from '@mui/material'
+import { Button, ButtonProps } from '@mui/material';
 
 const meta: Meta<ButtonProps> = {
   component: Button,
