+++ tsconfig.json
@@ -19,8 +19,6 @@
     ],
     "paths": {
       "@/*": ["./src/*"],
-      "@interface/*": ["./interface/*"],
-      "@utils/*": ["./utils/*"],
     },
   },
   "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
